baselnies

+++ utils/conv_type.py
@@ -128,7 +128,7 @@ class SubnetConvEdgePopup(nn.Conv2d):
                 weight_twin = torch.zeros_like(self.weight)
                 weight_twin = _init_weight(self.args, weight_twin)
                 ones = torch.ones(self.weight.size()).to(self.weight.device)
-                b = torch.bernoulli(ones * int(self.args.rerand_rate/2))
+                b = torch.bernoulli(ones * float(self.args.rerand_rate/2))
                 mask = GetSubnetEdgePopup.apply(self.clamped_scores, self.prune_rate)
                 t1 = self.weight.data * mask
                 t2 = self.weight.data * (1 - mask) * (1 - b)
